Improve email validation and specify email as unique index
The default email validation filter in PHP is fairly basic, this method will allow greater control on acceptable email addresses. For instance the 'single comma' character, while technically valid in the RFC 5322 specification it is generally considered an invalid character, by using regular expression we can decide if we want to accept this or not.